13079 - On the beach!
Input:
Input consists of a series of test cases. Each case begins with a line with the number N of buildings (1 ≤ N ≤ 100, 000). There follow N lines, each containing two integers representing the westernmost (Wi) and easternmost end (Ei) of each building, with Wi < Ei, measured in meters from the westernmost end of the beach. All these measures are integers between 0 and 109.
Input will end with a case without buildings, which should not be processed.

Output:
For each test case a line will be written with the minimum number of passageways that have to be built.
Passageways must be 1 meter wide and to be useful for a building they must be completely beneath it
as they pass through.

Sample Input:
4
1 4
6 15
2 10
12 20
2
1 4
4 8
2
1 4
3 8
0

Sample Output:
2
2
1