591 - Box of Bricks
Input:
The input consists of several data sets. Each set begins with a line containing the number n of stacks Bob has built. The next line contains n numbers, the heights hi of the n stacks. You may assume  $1 Ÿ\le n \leŸ 50$ and  $1 \leŸ h_i Ÿ\le 100$.
The total number of bricks will be divisible by the number of stacks. Thus, it is always possible to rearrange the bricks such that all stacks have the same height.
The input is terminated by a set starting with n = 0. This set should not be processed.

Output:
For each set, first print the number of the set, as shown in the sample output. Then print the line ``The minimum number of moves is k.'', where k is the minimum number of bricks that have to be moved in order to make all the stacks the same height.
Output a blank line after each set.

Sample Input:
6
5 2 4 1 7 5
0

Sample Output:
Set #1
The minimum number of moves is 5.