10062 - Tell me the frequencies!
Input:
Several lines of text are given as input. Each line of text is considered as a single input. Maximum
length of each line is 1000.

Output:
Print the ASCII value of the ASCII characters which are present and their frequency according to the
given format below. A blank line should separate each set of output. Print the ASCII characters in the
ascending order of their frequencies. If two characters are present the same time print the information
of the ASCII character with higher ASCII value first. Input is terminated by end of file.

Sample Input:
AAABBC
122333

Sample Output:
67 1
66 2
65 3

49 1
50 2
51 3