10191 - Longest Nap
Input:
The input will consist on an arbitrary number of test cases, each test case represents one day.
The first line of each set contains a positive integer s (not greater than 100) representing the
number of scheduled appointments during that day. In the next s lines there are the appointments in
the following format:
time1 time2 appointment
Where time1 represents the time which the appointment starts and time2 the time it ends. All
times will be in the ‘hh:mm’ format, time1 will always be strictly less than time2, they will be separated
by a single space and all times will be greater than or equal to 10:00 and less than or equal to 18:00.
So, your response must be in this interval as well (i.e. no nap can start before 10:00 and last after
18:00).
The appointment can be any sequence of characters, but will always be in the same line. You can
assume that no line will be longer than 255 characters, that 10 ≤ hh ≤ 18 and that 0 ≤ mm < 60.
You CAN’T assume, however,