10963 - The Swallowing Ground
Input:
The input begins with a single positive integer on a line by itself indicating the number of the cases
following, each of them as described below. This line is followed by a blank line, and there is also a
blank line between two consecutive inputs.
The first line contains the number of columns W of the ground to be considered. Then W lines
follow, specifying the boundary of the gap by columns: a line represents a column, and the order of
the lines follows the east-to-west order. Each line contains y1, the north-most row of the gap in this
column, and y2, the south-most row of the gap in this column. Row numbers are integers between -100
and 100 inclusive.
The input specifies a valid gap: the holes separate the ground into exactly a northern block and a
southern block. There are no holes embedded inside a block. There are no islands.
Note: The sample input below corresponds to the first diagram.

Output:
For each test case, the output must follow the descript