13078 - Feeding chicks
Input:
The input begins with the number of test cases to be processed. Each case begins with three numbers,
r, c and n meaning, respectively, the size of the chicken coop in the north-south direction, the size
in the east-west direction and the number of chicks (1 ≤ r, c ≤ 50; 0 ≤ n ≤ 500). The next n lines
contain the chicks information. The first number, v, means the position in the north-south direction
(1 ≤ v ≤ r), the second, h, the position in the east-west direction (1 ≤ h ≤ c), next it is the start
direction of the chick (N, S, E, W) and finally the maximum number of steps before falling sleep (at
least one).

Output:
For each test case r lines are written. In each line c values separated by a blank are written, showing
the number of grains needed in each point. After each test case write three scripts (---).

Sample Input:
2
7 8 3
1 2 E 2
6 3 N 25
4 6 E 21
4 3 2
1 1 N 3
4 2 W 3

Sample Output:
0 1 1 0 0 0 0 0
0 0 1 1 1 0 0 0
0 0 0 1 1